♻️ Move dummy hook into redux store
